feat: macros calculations when getting meal plans

+++ src/NutriWise.Application/Nutrition/NutritionPlan.cs
@@ -2,8 +2,8 @@
 
 public class NutritionPlan
 {
-	public double Calories { get; set; }
-	public double ProteinGrams { get; set; }
-	public double FatGrams { get; set; }
-	public double CarbGrams { get; set; }
+	public int Calories { get; set; }
+	public int ProteinGrams { get; set; }
+	public int FatGrams { get; set; }
+	public int CarbGrams { get; set; }
 }
